[DELTA-OSS-EXTERNAL] Make JsonUtils.mapper lazy to show a better error
Currently if failing to initialize `JsonUtils.mapper` (likely `jackson-module-scala` incompatibility issue), class `JsonUtils$` will not be loaded. Because this is a class loading error, the root cause of initialization failure will be shown only once by JVM, and the following calls that touch `JsonUtils` will fail without a cause. This makes users hard to find out the root cause in their logs.

This PR makes JsonUtils.mapper lazy so that the initialization failure will be shown every time `JsonUtils` is touched.

Closes delta-io/delta#552

Signed-off-by: Rahul Mahadev <rahul.mahadev@databricks.com>

Author: Shixiong Zhu <zsxwing@gmail.com>

#14292 is resolved by rahulsmahadev/msqzx64j.

GitOrigin-RevId: de71894fe398d0eeaa210ed8e837dbee4bf094ef