[SC-45268] Save comment when CTAS a delta table with comment
## What changes were proposed in this pull request?

When create a new delta table with a comment by CTAS, the metadata of the new table
misses the comment.

Replace delta table command updates the comment in a update metadata request, so the metadta is
correct though the comment is not set in the initial metadata.

To fix the issue, comment is set to metadata  uring metadata creation.

## How was this patch tested?

The new unit tests.

Author: Alex liu <alex_liu68@yahoo.com>

Author: Alex liu <alex_liu68@yahoo.com>
Author: Burak Yavuz <brkyvz@gmail.com>

#14249 is resolved by alexoss68/SC-45268-master.

GitOrigin-RevId: e6c56b08a7ec55be63e05f888b8d968f62cca6d4